+++ src/txmath1.cpp
@@ -1866,7 +1866,7 @@ auto Math::convert_char_seq(MathElt W) -> MathElt {
     if (f == 1) B.push_back(' ');
     res = new Xml(Istring(B));
     res = new Xml(cst_mi, res);
-    if (f > 1 && spec) res->add_att(cst_mathvariant, name_positions(cstf_normal + f));
+    if (f > 1 && spec) res->add_att(cst_mathvariant, name_positions(long(cstf_normal) + long(f)));
     return MathElt(res, mt_flag_small);
 }
 
@@ -1889,6 +1889,6 @@ auto Math::convert_char_iseq(MathElt W, bool multiple) -> MathElt {
         }
     Xml *res = new Xml(Istring(B));
     res      = new Xml(cst_mn, res);
-    if (f > 1) res->add_att(cst_mathvariant, name_positions(cstf_normal + f));
+    if (f > 1) res->add_att(cst_mathvariant, name_positions(long(cstf_normal) + long(f)));
     return MathElt(res, mt_flag_small);
 }
